server/requirements.txt
django
djangorestframework
psycopg2-binary==2.8.5
pytest-django==3.9.0
pytest==5.4.1
gunicorn
django-cors-headers